feat: add svelte kit app (#589)
