+++ examples/svelte-kit/.eslintrc.cjs
