+++ examples/svelte-kit/.gitignore
