+++ examples/svelte-kit/README.md
