+++ examples/svelte-kit/src/app.d.ts
