+++ examples/svelte-kit/src/components/toolbar.svelte
