+++ examples/svelte-kit/src/routes/+page.svelte
@@ -1,0 +1,16 @@
+<script>
+	import { routesData } from '../../../../shared/src/routes';
+
+	const routes = routesData.sort((a, b) => a.label.localeCompare(b.label));
+</script>
+
+<div class="index-nav">
+	<h2>Zag.js + React</h2>
+	<ul>
+		{#each routes as route}
+			<li>
+				<a href={route.path}>{route.label}</a>
+			</li>
+		{/each}
+	</ul>
+</div>
