+++ examples/svelte-kit/static/favicon.png
