--- examples/svelte-ts/svelte.config.js
@@ -1,7 +1,0 @@
-import { vitePreprocess } from "@sveltejs/vite-plugin-svelte"
-
-const config = {
-  preprocess: vitePreprocess(),
-}
-
-export default config
