--- examples/svelte-ts/vite.config.ts
@@ -1,6 +1,0 @@
-import { defineConfig } from "vite"
-import { svelte } from "@sveltejs/vite-plugin-svelte"
-
-export default defineConfig({
-  plugins: [svelte()],
-})
